+++ packages/pipeline/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tangleai/pipeline
 
+## 0.20.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @tangleai/core@0.20.1
+  - @tangleai/memory@0.20.1
+
 ## 0.20.0
 
 ### Minor Changes